RF: make compatible also with recent 0.7.0
But having done it not sure if I we want to keep compatibility with
previuos versions since it is not only that API changed, but also their
annotation. At the least  "type" is now a non-specific "suffix", and
"modality" became a "datatype".  Whenever I am ok with "datatype",
loosing "type" is probably suboptimal although makes sense within
BIDS itself.

But having those changes in mind, it might be better to just demand
0.7.0 and remove compatibility layer.

Also I think it is time to start encoding some kind of "version" of
PyBIDS convention on top of what version of BIDS actual dataset
conforms with.  Since BIDS does not necessarily always gives corresponding
consistent namings aligned with PyBIDS I feel that we could still
have similar renamings in the future