MNT: nipype_workshop_dataset: Update save() call for DataLad 0.12
In 0.12.x, `datalad save` no longer accepts --super-datasets.  In a
blind to make the example compatible with the latest DataLad, use the
--dataset=^ shortcut instead.

Note that there are other spots in the example that should be updated
for the 0.12.x (e.g., use of `datalad add`).  This is just a minimal
attempt to get it passing.

Re: https://github.com/datalad/datalad-neuroimaging/pull/74#issuecomment-580340491