TST: Add appveyor setup
This is supposed to replace Travis and comes with a build on Ubuntu,
MacOS and Windows respectively. However, we didn't have any mac/win
builds before, so those builds are simply meant to provide the setup for
future work towards supporting these platforms. Therefore also skip all
tests but test_register on OSX and Windows for now.