BF: nifti1 - decode "bytes" in the metadata record
While trying out metalad using nifti1 extractor as the one which needs "sparse"
access to actual file content, it was crashing due to

    TypeError: Object of type bytes is not JSON serializable

which were due to the fact that some fields, e.g.

    *(Pdb) p str(header['descrip'])
    "b'FreeSurfer Aug 11 2009'"

are byte numpy arrays... I decided to just to the final tune up for all values
regardless where they came from - I do not envision a use case where they should
be a byte string.

If it would be the case -- metalad would need to be the one to handle that
somehow I guess.